refactor: move some functions from `Animal` to `AgeableMob`
