+++ steel-core/src/entity/entities/mobs/passive/cow/tests/mod.rs
@@ -10,7 +10,7 @@ use steel_registry::{
 use steel_utils::types::InteractionHand;
 
 use crate::entity::damage::DamageSource;
-use crate::entity::{Animal, Entity, LivingEntity, Mob};
+use crate::entity::{Entity, LivingEntity, Mob};
 use crate::test_support::{TestPlayerBuilder, fresh_test_world};
 
 use super::*;
