+++ steel-core/src/entity/entities/mobs/passive/pig/mod.rs
@@ -408,20 +408,6 @@ impl AgeableMob for PigEntity {
     fn set_synced_baby(&self, baby: bool) {
         self.entity_data.lock().ageable_mob_mut().baby.set(baby);
     }
-}
-
-impl Animal for PigEntity {
-    fn animal_base(&self) -> &AnimalBase {
-        &self.animal_base
-    }
-
-    fn is_food(&self, item_stack: &ItemStack) -> bool {
-        PigEntity::is_food(item_stack)
-    }
-
-    fn play_eating_sound(&self) {
-        self.play_sound(self.current_sound_set().eat_sound, 1.0, 1.0);
-    }
 
     fn breed_variant_key(&self) -> Option<&Identifier> {
         Some(&self.variant().key)
@@ -431,7 +417,7 @@ impl Animal for PigEntity {
         self.set_variant_by_key(key)
     }
 
-    fn initialize_breed_offspring(&self, partner: &dyn Animal, offspring: &dyn Animal) {
+    fn initialize_breed_offspring(&self, partner: &dyn AgeableMob, offspring: &dyn AgeableMob) {
         let use_self_variant = rand::random::<bool>();
         let variant_key = if use_self_variant {
             self.breed_variant_key()
@@ -445,6 +431,20 @@ impl Animal for PigEntity {
         if !offspring.set_breed_variant_key(variant_key) {
             log::error!("pig offspring could not inherit breeding variant {variant_key}");
         }
+    }
+}
+
+impl Animal for PigEntity {
+    fn animal_base(&self) -> &AnimalBase {
+        &self.animal_base
+    }
+
+    fn is_food(&self, item_stack: &ItemStack) -> bool {
+        PigEntity::is_food(item_stack)
+    }
+
+    fn play_eating_sound(&self) {
+        self.play_sound(self.current_sound_set().eat_sound, 1.0, 1.0);
     }
 }
